Web Browser Module
A simple python script used to open the location of any place from
windows powershell. Just Press windows+r and then type the name of the
place whose location you want to search, and press enter. It will open
the location of the place in your default browser in a new tab.